refactor: 重构灵动岛主题切换接口
